Fixed warning in mtsStateIndex.h.
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@2076 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/include/cisstMultiTask/mtsStateIndex.h
@@ -117,7 +117,7 @@ public:
 	/*! Overloaded operator */
 	mtsStateIndex& operator-=(int number) {
 		TimeTicks -= number;
-		TimeIndex = TimeTicks % BufferLength;
+		TimeIndex = static_cast<int>(TimeTicks % BufferLength);
 		return *this;
 	}
